refactor: split remaining ≥260 widget and screen hosts (#4582)
Slice D extracts sibling libraries along existing seams so gated app/lib
trees stay at or under 260 lines without changing player-visible behavior.
Naval fleet-marker and mission flow entrypoints keep the same signatures.
