+++ app/lib/features/game/flame/map_state/game_map_area_province_action_states.dart
@@ -1,5 +1,4 @@
 import 'package:colonizethis_data/colonizethis_data.dart';
-
 import 'package:colonizethis_map/colonizethis_map.dart';
 import 'package:colonizethis_models/colonizethis_models.dart' as ct_models;
