+++ app/lib/features/game/widgets/unit_orders/naval_mission_flow.dart
@@ -1,7 +1,6 @@
 // Naval mission assign + map fleet-marker routing (Refs #4213, #4343, #4448).
 // SPEC/program/app-ui-wiring.md.
 
-import 'package:colonizethis_app_l10n/l10n/l10n.dart';
 import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_orders/colonizethis_orders.dart'
@@ -9,13 +8,14 @@ import 'package:colonizethis_orders/colonizethis_orders.dart'
 import 'package:colonizethis_world/colonizethis_world.dart';
 import 'package:flutter/material.dart';
 
-import '../../../../config/ui_screen_ids.dart';
-import '../province_overlay/province_sea_zone_detail_overlay_designation.dart';
-import '../panels/tree_builders/fleet_mission_label.dart';
 import 'home_fleet_detach_then_sail_flow.dart';
-import 'move_fleet_dialog.dart';
 import 'naval_mission_flow_support.dart';
+import 'naval_mission_flow_types.dart';
 import 'naval_mission_menu_dialog.dart';
+import 'naval_mission_move_dialog.dart';
+
+export 'naval_mission_flow_types.dart';
+export 'naval_mission_move_dialog.dart';
 
 /// Map fleet-marker tap: pick fleet when stacked, then route to the legal action.
 ///
@@ -217,81 +217,3 @@ Future<void> showNavalMissionFlow({
       }
   }
 }
-
-/// Local `showDialog` for [MoveFleetDialog] (panel Move, marker in-port, Sail).
-Future<bool?> showMoveFleetDialogForFleet({
-  required BuildContext context,
-  required Game game,
-  required MapTopology topology,
-  required String humanPlayerId,
-  required Fleet fleet,
-  required AppEventBus bus,
-  PlayerView? playerView,
-}) {
-  final resolvedPlayerView =
-      playerView ?? buildPlayerView(game, topology, humanPlayerId);
-  return showDialog<bool>(
-    context: context,
-    builder: (ctx) => MoveFleetDialog(
-      game: game,
-      topology: topology,
-      humanPlayerId: humanPlayerId,
-      fleet: fleet,
-      bus: bus,
-      playerView: resolvedPlayerView,
-    ),
-  );
-}
-
-/// Menu selection for [showNavalMissionFlow].
-sealed class NavalMissionMenuChoice {
-  const NavalMissionMenuChoice();
-}
-
-final class NavalMissionMenuChoiceMission extends NavalMissionMenuChoice {
-  const NavalMissionMenuChoiceMission(this.mission);
-  final FleetMission mission;
-}
-
-final class NavalMissionMenuChoiceCancelPending extends NavalMissionMenuChoice {
-  const NavalMissionMenuChoiceCancelPending();
-}
-
-/// Opens [MoveFleetDialog] from `DLG31001` Sail / Move (Refs #4343).
-final class NavalMissionMenuChoiceSail extends NavalMissionMenuChoice {
-  const NavalMissionMenuChoiceSail();
-}
-
-/// Screen ids for naval mission dialog hosts (Refs #4213).
-abstract final class NavalMissionDialogIds {
-  static const menuDialog = UiScreenIds.navalMissionMenuDialog;
-  static const targetDialog = UiScreenIds.navalMissionTargetDialog;
-  static const fleetPickerDialog = UiScreenIds.navalMissionFleetPickerDialog;
-}
-
-String navalMissionMenuLabel(FleetMission mission) =>
-    fleetMissionDisplayLabel(mission);
-
-/// One-line player-facing effect summary for [mission] (Refs #4295).
-String navalMissionEffectLine(AppLocalizations l10n, FleetMission mission) {
-  return switch (mission) {
-    FleetMission.patrol => l10n.naval_mission_effect_patrol,
-    FleetMission.defend => l10n.naval_mission_effect_defend,
-    FleetMission.blockade => l10n.naval_mission_effect_blockade,
-    FleetMission.beachhead => l10n.naval_mission_effect_beachhead,
-    FleetMission.none => '',
-  };
-}
-
-/// Target-picker caption for Blockade / Beachhead (Refs #4295).
-String? navalMissionTargetCaption(AppLocalizations l10n, FleetMission mission) {
-  return switch (mission) {
-    FleetMission.blockade => l10n.naval_mission_targetCaption_blockade,
-    FleetMission.beachhead => l10n.naval_mission_targetCaption_beachhead,
-    _ => null,
-  };
-}
-
-/// Whether [provinceId] is any faction's capital province (Blockade assign).
-bool navalMissionBlockadeTargetIsCapital(Game game, String provinceId) =>
-    provinceOverlayIsCapital(game, provinceId);
